More tidy up

+++ src/BackTester/ohlcdatapoint.h
@@ -4,17 +4,17 @@
 #include <string>
 
 struct OHLCDataPoint {
-    OHLCDataPoint(const std::string& date, int open, int high, int low, int close, int volume, int adjClose)
+    OHLCDataPoint(const std::string& date, double open, double high, double low, double close, double volume, double adjClose)
         : date(date), open(open), high(high), low(low), close(close), volume(volume), adjClose(adjClose)
     {}
 
     std::string date;
-    int open;
-    int high;
-    int low;
-    int close;
-    int volume;
-    int adjClose;
+    double open;
+    double high;
+    double low;
+    double close;
+    double volume;
+    double adjClose;
 };
 
 #endif // OHLCDATAPOINT_H
